+++ src/Type/UniqueGiftInfo.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Vjik\TelegramBot\Api\Type;
 
+use DateTimeImmutable;
+
 /**
  * @see https://core.telegram.org/bots/api#uniquegiftinfo
  *
@@ -16,5 +18,7 @@ final readonly class UniqueGiftInfo
         public string $origin,
         public ?string $ownedGiftId = null,
         public ?int $transferStarCount = null,
+        public ?int $lastResaleStarCount = null,
+        public ?DateTimeImmutable $nextTransferDate = null,
     ) {}
 }